Add current/hover values to the plot legend
